refactoring lesson
