+++ src/pages-content/chess-lesson/chess-lesson-board-game/constants/game.tsx
