+++ src/pages-content/chess-lesson/chess-lesson-board-game/piece-white-knight/knight-piece.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useDrag } from 'react-dnd'
 
 import pieces from '../../../../chess-board-game/pieces/chesspieces'
-import { ItemTypes } from './constants/ItemTypes'
+import { ItemTypes } from '../constants/ItemTypes'
 export default function Knight() {
   const [{ isDragging }, drag] = useDrag(() => ({
     type: ItemTypes.KNIGHT,
